admin blog delete function modify

+++ src/app/modules/admin/admin.service.ts
@@ -1,4 +1,5 @@
 import { UserRegister } from '../auth/auth.model';
+import { currentUserEmail } from '../auth/auth.utils';
 import { BlogPost } from '../blogs/blog.model';
 
 const blockUserByAdminIntoDb = async (userId: string) => {
@@ -11,6 +12,12 @@ const blockUserByAdminIntoDb = async (userId: string) => {
 };
 
 const deleteBlogByAdminIntoDb = async (id: string) => {
+  const receivedEmail = currentUserEmail;
+  const findUser = await UserRegister.findOne({ email: receivedEmail });
+  console.log(findUser)
+  if(findUser?.role!=='admin'){
+    throw new Error('You are not an admin !')
+  }
   const result = await BlogPost.findByIdAndDelete(id);
   return result;
 };
